feat: create module to get data from api - ease the switch of data loader format
